Separate comments DBs and classes, because it makes no sense to make comments into points.
https://github.com/Tour-de-Force/btc-models/issues/17
